Added hlmer2() function
Added the hlmer2() function, which allows users to specify HLM models
with a level-1 regression equation and a list of level-2 regression
equations. See the ?hlmer2 help file for more information and for
examples.

Also added the center_data() function that will create centered
versions of all numeric variables in a data set (and retain the
original versions, too).